+++ src/Http/Controllers/UserController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tychovbh\Mvc\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class UserController extends Controller
+{
+    //
+}