+++ src/MvcServiceProvider.php
@@ -3,6 +3,7 @@
 namespace Tychovbh\Mvc;
 
 use Illuminate\Support\ServiceProvider;
+use Laravelista\LumenVendorPublish\VendorPublishCommand;
 use Tychovbh\Mvc\Console\Commands\MvcRepository;
 use Tychovbh\Mvc\Console\Commands\MvcController;
 use Tychovbh\Mvc\Console\Commands\MvcRequest;
@@ -17,24 +18,30 @@ class MvcServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        if ($this->app->runningInConsole()) {
-            $this->commands([
-                MvcRepository::class,
-                MvcController::class,
-                MvcRequest::class
-            ]);
+        if (!$this->app->runningInConsole()) {
+            return;
         }
 
+        $this->commands([
+            MvcRepository::class,
+            MvcController::class,
+            MvcRequest::class,
+        ]);
+
         $source = __DIR__ . '/../config/messages.php';
-        if (is_application() === 'laravel') {
-            $this->publishes([$source => config_path('messages.php')], 'laravel-mvc');
-        }
+        $this->publishes([$source => config_path('messages.php')], 'laravel-mvc-config');
+        $this->publishes([
+            __DIR__.'/../database/migrations' => database_path('migrations'),
+        ], 'laravel-mvc-migrations');
 
         if (is_application() === 'lumen') {
             $this->app->configure('messages');
             $this->app->register(\Urameshibr\Providers\FormRequestServiceProvider::class);
             $this->app->routeMiddleware([
                 'validate' => ValidateMiddleware::class
+            ]);
+            $this->commands([
+                VendorPublishCommand::class
             ]);
         }
     }