+++ src/Repositories/UserRepository.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Tychovbh\Mvc\Repositories;
+
+use Tychovbh\Mvc\Repositories\Repository;
+use Tychovbh\Mvc\Repositories\AbstractRepository;
+
+class UserRepository extends AbstractRepository implements Repository
+{
+    //
+}