+++ tests/TestCase.php
@@ -6,6 +6,7 @@ namespace Tychovbh\Tests\Mvc;
 
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Tychovbh\Mvc\MvcServiceProvider;
+use Tychovbh\Tests\Mvc\App\TestUserController;
 
 class TestCase extends BaseTestCase
 {
@@ -15,8 +16,9 @@ class TestCase extends BaseTestCase
     public function setUp(): void
     {
         parent::setUp();
-        $this->withFactories(__DIR__.'/../database/factories');
+        $this->withFactories(__DIR__ . '/database/factories');
         $this->artisan('migrate', ['--database' => 'testing']);
+        $this->loadMigrationsFrom(__DIR__ . '/database/migrations');
         $this->loadMigrationsFrom(__DIR__ . '/../database/migrations');
     }
 
@@ -28,8 +30,15 @@ class TestCase extends BaseTestCase
      */
     protected function getEnvironmentSetUp($app)
     {
-        $app['config']->set('database.default', 'testing');
         $app['config']->set('env', 'testing');
+        $app['config']->set('database.default', 'testbench');
+        $app['config']->set('database.connections.testbench', [
+            'driver'   => 'sqlite',
+            'database' => ':memory:',
+            'prefix'   => '',
+        ]);
+
+        $app['router']->get('users', TestUserController::class . '@index')->name('users.index');
     }
 
     /**
@@ -39,24 +48,5 @@ class TestCase extends BaseTestCase
     protected function getPackageProviders($app)
     {
         return [MvcServiceProvider::class];
-    }
-
-    /** @test */
-    public function itRunsTheMigrations()
-    {
-        $users = \DB::table('test_users')->where('id', '=', 1)->first();
-        $this->assertEquals('hello@orchestraplatform.com', $users->email);
-        $this->assertTrue(\Hash::check('123', $users->password));
-        $columns = \Schema::getColumnListing('test_users');
-        $this->assertEquals([
-            'id',
-            'email',
-            'password',
-            'firstname',
-            'surname',
-            'hidden',
-            'created_at',
-            'updated_at',
-        ], $columns);
     }
 }