+++ tests/app/TestUser.php
@@ -5,4 +5,5 @@ namespace Tychovbh\Tests\Mvc\App;
 class TestUser extends \Illuminate\Database\Eloquent\Model
 {
     protected $fillable = ['email', 'password', 'firstname', 'surname'];
+    protected $hidden = ['password'];
 }